web_cli_gemini/requirements.txt
Flask
google-genai 
python-dotenv